pyproject.toml
[tool.poetry]
name = "python-bot"
version = "0.1.0"
description = ""
authors = ["rhodinemma <rhodinemma10@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
tweepy = "^4.14.0"
fastapi = "^0.110.0"
uvicorn = "^0.29.0"
selenium = "^4.18.1"
instaloader = "^4.11"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
